Simplify error reporting.
Thinking about this more, returning the number of bytes along with
a failure to indicate a partial success makes user code more complex.
So I'm now thinking it's better to just have the interface be a single
result which either succeeds or fails.

This also removes dependence on the `$errno` type, as I/O types don't
need the majority of the error values.