Add a polling interface for streams. (#20)
The new wasi-poll API has a `pollable` type. Add subscribe methods to
the stream types to allow polling for stream events.

This adds a dependency on wasi-io, and is currently implemented in the
repository in a low-tech way, which allows the generated documentation
to work, though it's likely to evolve.