Make input-stream and output-stream always non-blocking. (#24)
Specify that reads and writes are always non-blocking, to the extent
practical on underlying platforms. This is closer to how the future
builtin `stream` type will work, and should be more convenient for
users doing async I/O.

To do blocking reads or writes, use a `poll_oneoff` to wait if an initial
read or write didn't transmit any data. The first example in the
README.md gives an example of how to do this.

It also eliminates the need for a `would-block` error, as `would-block`
can be translated to a result of 0 bytes being transmitted. As such, this
is an alternative to #23.