Add explicit blocking functions.
With #24 making streams always non-blocking, add functions to make it
easy for users to use streams in blocking ways. This way, callers chose
whether to block or not by calling the `blocking`- or regular
non-blocking functions.

Along with #24, this switches wasi-io streams from having blockingness as a
dynamic flag associated with a stream, to being an API choice that callers
make in their choice of which function to use.