Move error resource to its own interface
Two reasons for this design change:

it is confusing having one type named stream/stream-error and another named stream/error.
this error resource seems useful outside of just streams.
Therefore, we are moving it to a separate interface error in the same package. There are no functional changes to this type.

The doc comments are now more general.